backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "huggingface>=0.0.1",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "pgvector>=0.3.6",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pwdlib>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "youtube-transcript-api==1.2.3",
]
